monexamine: move pet_menu out of game.cpp
A very basic refactoring of game.cpp and monexamine.cpp: move pet_meu
out of game.cpp in preparation for cleaning it up some.
