+++ src/monexamine.h
@@ -6,6 +6,7 @@ class monster;
 
 namespace monexamine
 {
+bool pet_menu( monster *z );
 /*
 *Manages the milking and milking cool down of monsters.
 *Milked item uses starting_ammo, where ammo type is the milked item
